BACKPORT: tcp: Add reuseport_migrate_sock() to select a new listener.
[ Upstream commit 1cd62c21572c1df6e7090ea4cabf4cf509616dbb ]

reuseport_migrate_sock() does the same check done in
reuseport_listen_stop_sock(). If the reuseport group is capable of
migration, reuseport_migrate_sock() selects a new listener by the child
socket hash and increments the listener's sk_refcnt beforehand. Thus, if we
fail in the migration, we have to decrement it later.

We will support migration by eBPF in the later commits.

[Mathias Gluszczynski: Use Linux 4.9's atomic socket reference count
when retaining the selected listener.]

Signed-off-by: Kuniyuki Iwashima <kuniyu@amazon.co.jp>
Signed-off-by: Martin KaFai Lau <kafai@fb.com>
Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>
Reviewed-by: Eric Dumazet <edumazet@google.com>
Link: https://lore.kernel.org/bpf/20210612123224.12525-5-kuniyu@amazon.co.jp
Signed-off-by: Mathias Gluszczynski <admin@krazey.de>