BACKPORT: tcp: Add stats for socket migration.
[ Upstream commit 55d444b310c64b084dcc62ba3e4dc3862269fb96 ]

This commit adds two stats for the socket migration feature to evaluate the
effectiveness: LINUX_MIB_TCPMIGRATEREQ(SUCCESS|FAILURE).

If the migration fails because of the own_req race in receiving ACK and
sending SYN+ACK paths, we do not increment the failure stat. Then another
CPU is responsible for the req.

[Mathias Gluszczynski: Append the counters to Linux 4.9's current MIB
table and adapt the accounting to its timer callbacks, ehash insertion,
and atomic request references.]

Link: https://lore.kernel.org/bpf/CAK6E8=cgFKuGecTzSCSQ8z3YJ_163C0uwO9yRvfDSE7vOe9mJA@mail.gmail.com/
Suggested-by: Yuchung Cheng <ycheng@google.com>
Signed-off-by: Kuniyuki Iwashima <kuniyu@amazon.co.jp>
Acked-by: Yuchung Cheng <ycheng@google.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Mathias Gluszczynski <admin@krazey.de>