BACKPORT: bpf: Refactor perf_event_set_bpf_prog() input
[ Upstream commit 652c1b17b85b9c195978c051aa283027529db1fe ]

Make internal perf_event_set_bpf_prog() use struct bpf_prog pointer as an
input argument, which makes it easier to re-use for other internal uses
(coming up for BPF link in the next patch). BPF program FD is not as
convenient and in some cases it's not available. So switch to struct bpf_prog,
move out refcounting outside and let caller do bpf_prog_put() in case of an
error. This follows the approach of most of the other BPF internal functions.

[Mathias Gluszczynski: Refactor the older perf ioctl and tracepoint
paths. Leave newer precise-IP and kprobe-override checks for their own
prerequisite commits because this 4.9 tree does not contain them.]

Signed-off-by: Andrii Nakryiko <andrii@kernel.org>
Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>
Acked-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Acked-by: Yonghong Song <yhs@fb.com>
Link: https://lore.kernel.org/bpf/20210815070609.987780-4-andrii@kernel.org
Signed-off-by: Mathias Gluszczynski <admin@krazey.de>