BACKPORT: net: Rename ->stream_memory_read to ->sock_is_readable
[ Upstream commit 7b50ecfcc6cdfe87488576bc3ed443dc8d083b90 ]

The proto ops ->stream_memory_read() is currently only used
by TCP to check whether psock queue is empty or not. We need
to rename it before reusing it for non-TCP protocols, and
adjust the exsiting users accordingly.

[Mathias Gluszczynski: Rename the TCP and sockmap users present in
Linux 4.9. Omit TLS users because this tree lacks their corresponding
readability callback.]

Signed-off-by: Cong Wang <cong.wang@bytedance.com>
Signed-off-by: Alexei Starovoitov <ast@kernel.org>
Link: https://lore.kernel.org/bpf/20211008203306.37525-2-xiyou.wangcong@gmail.com
Signed-off-by: Mathias Gluszczynski <admin@krazey.de>