BACKPORT: net: Implement ->sock_is_readable() for UDP and AF_UNIX
[ Upstream commit af493388950b6ea3a86f860cfaffab137e024fc8 ]

Yucong noticed we can't poll() sockets in sockmap even
when they are the destination sockets of redirections.
This is because we never poll any psock queues in ->poll(),
except for TCP. With ->sock_is_readable() now we can
overwrite >sock_is_readable(), invoke and implement it for
both UDP and AF_UNIX sockets.

[Mathias Gluszczynski: Apply the UDP portion to Linux 4.9 and omit
AF_UNIX because this backport does not add UNIX sockmap support.]

Reported-by: Yucong Sun <sunyucong@gmail.com>
Signed-off-by: Cong Wang <cong.wang@bytedance.com>
Signed-off-by: Alexei Starovoitov <ast@kernel.org>
Link: https://lore.kernel.org/bpf/20211008203306.37525-4-xiyou.wangcong@gmail.com
Signed-off-by: Mathias Gluszczynski <admin@krazey.de>