BACKPORT: bpf: Mark PTR_TO_FUNC register initially with zero offset
[ Upstream commit d400a6cf1c8a57cdf10f35220ead3284320d85ff ]

Similar as with other pointer types where we use ldimm64, clear the
register
content to zero first, and then populate the PTR_TO_FUNC type and
subprogno
number. Currently this is not done, and leads to reuse of stale register
tracking data.

Given for special ldimm64 cases we always clear the register offset,
make it
common for all cases, so it won't be forgotten in future.

[Mathias Gluszczynski: Adapt the common initialization point to the
special ldimm64 forms supported by this 4.9 verifier.]

Fixes: 69c087ba6225 ("bpf: Add bpf_for_each_map_elem() helper")
Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>
Acked-by: John Fastabend <john.fastabend@gmail.com>
Acked-by: Alexei Starovoitov <ast@kernel.org>
Signed-off-by: Mathias Gluszczynski <admin@krazey.de>