BACKPORT: bpf: Fix JIT blinding with pointers to BPF subprogs
The combination of JIT blinding and pointers to BPF subprogs causes:

  BUG: unable to handle page fault for address: 0000000100000001
  #PF: supervisor instruction fetch in kernel mode
  #PF: error_code(0x0010) - not-present page
  RIP: 0010:0x100000001

The JIT blinding logic did not recognize that ld_imm64 with an address
of a BPF subprogram is a special instruction and proceeded to randomize
it. By itself it would not have been an issue, but jit_subprogs() relies
on a two-step process to JIT all subprograms and then JIT them
again when addresses of all subprograms are known. Blinding in the
first JIT phase caused the second JIT to miss adjustment of the special
ld_imm64.

Fix this issue by ignoring special ld_imm64 instructions that do not
have user-controlled constants and should not be blinded.

Fixes: 69c087ba6225 ("bpf: Add bpf_for_each_map_elem() helper")
Reported-by: Andrii Nakryiko <andrii@kernel.org>
Signed-off-by: Alexei Starovoitov <ast@kernel.org>
Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>
Acked-by: Andrii Nakryiko <andrii@kernel.org>
Acked-by: Martin KaFai Lau <kafai@fb.com>
Link: https://lore.kernel.org/bpf/20220513011025.13344-1-alexei.starovoitov@gmail.com
(cherry picked from commit 4b6313cf99b0d51b49aeaea98ec76ca8161ecb80)
Signed-off-by: Mathias Gluszczynski <admin@krazey.de>