BACKPORT: ANDROID: userfaultfd: add MMAP_TRYLOCK mode for COPY/ZEROPAGE
In case mmap_lock is contended, it is possible that userspace can spend
time performing other tasks rather than waiting in uninterruptible-sleep
state for the lock to become available. Even if no other task is
available, it is better to yield or sleep rather than adding contention
to already contended lock.

We introduce MMAP_TRYLOCK mode so that when possible, userspace can
request to use mmap_read_trylock(), returning -EAGAIN if and when it
fails.

Bug: 320478828
Change-Id: I2d196fd317e054af03dbd35ac1b0c7634cb370dc
Signed-off-by: Lokesh Gidra <lokeshgidra@google.com>
Signed-off-by: Cyber Knight <cyberknight755@gmail.com>

[Mathias Gluszczynski: Keep fault_env and use BUILD_BUG_ON for the
mode-bit assertion instead of importing a new static_assert interface.]

Upstream-commit: 954fa80da1145409735c353cbe8e44e7cf713236
Signed-off-by: Mathias Gluszczynski <admin@krazey.de>