BACKPORT: ANDROID: userfaultfd: abort on mmap lock contention
Original subject: ANDROID: userfaultfd: abort uffdio ops if mmap_lock is
contended

Check if the mmap_lock is contended when looping over the pages that are
requested to be filled. When it is observed, we rely on the already
existing mechanism to return bytes copied/filled and -EAGAIN as error.

This helps by avoiding contention of mmap_lock for long running
userfaultfd operations. The userspace can perform other tasks before
retrying the operation for the remaining pages.

Bug: 320478828
Change-Id: I6d485fd03c96a826956ee3962e58058be3cf81c1
Signed-off-by: Lokesh Gidra <lokeshgidra@google.com>
Signed-off-by: Cyber Knight <cyberknight755@gmail.com>

Upstream-commit: dbf7b9226fbb6a812ba1348ce7e1ee55b55d6ec4
Signed-off-by: Mathias Gluszczynski <admin@krazey.de>